Implement exec collector
So that salmon can now run the same command periodically and ensure that
its exit code is as expected. Later on we can also add other things to
check, like if stdout and/or stderr contain or don't contain certian
substrings, but for now only exit code can be checked.